pyproject.toml
[project]
name = "polymarket-arbitrage"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "websocket-client>=1.6.0",
    "requests>=2.31.0",
    "numpy>=1.24.0",
    "fastapi>=0.110.0",
    "uvicorn>=0.27.0",
    "pyyaml>=6.0",
    "anthropic>=0.40.0",
    "google-genai>=1.0.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-cov"]
